requirements.txt
openai
pyyaml
pydantic
faiss-cpu
numpy
requests
gradio
python-dotenv
pymongo
python-docx
pypdf
